A2L reform
In light of one of the most definitive authorities of Nissan/Renault ROMs becoming available, the definitions have been updated to remove experimental definitions. Also, in an effort to assist with cross-referencing between the A2L file and RR definitions, the short A2L table names are being incorporated into the table templates.